backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==21.2.0
tensorflow-cpu==2.15.0
numpy